LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-invalid-header-expected.txt

CONSOLE MESSAGE: Error while parsing the 'sandbox' Content Security Policy directive: 'allowScript' is an invalid sandbox flag.
